chore(shared): AGP9 KMP 플러그인 마이그레이션 + CMP 1.11.1 / kotlinx-datetime 0.7.1 정합
- shared: com.android.library → com.android.kotlin.multiplatform.library (AGP9 필수)
- Compose Multiplatform 1.7.0 → 1.11.1 (Kotlin 2.3.21 호환), Decompose 3.0.0 → 3.5.0
- CMP 1.11 미지원인 iosX64(Intel 시뮬레이터) 타깃 제거
- material-icons-core 1.7.3 직접 의존 (CMP 1.8+에서 전이 의존 제거됨)
- kotlinx-datetime 0.7.1 통일: Clock/Instant → kotlin.time 마이그레이션
- DayOfWeek java.time typealias 제거 대응 (getDisplayName 호환 확장, isoDayNumber)
- Koin DataModule: TodoRepositoryImpl 신규 파라미터(localUserConfigDataSource) 반영
- .gitignore: KMP 브랜치에서는 iosApp/ 추적 유지

+++ core/designsystem/src/main/java/com/tgyuu/designsystem/component/calendar/CalendarDate.kt
@@ -27,7 +27,8 @@ fun getWeekStart(date: LocalDate, startFromMonday: Boolean = false): LocalDate {
 fun weeksBetween(from: LocalDate, to: LocalDate, startFromMonday: Boolean = false): Int {
     val fromWeekStart = getWeekStart(from, startFromMonday)
     val toWeekStart = getWeekStart(to, startFromMonday)
-    return (toWeekStart.toEpochDays() - fromWeekStart.toEpochDays()) / 7
+    // kotlinx-datetime 0.7부터 toEpochDays()가 Long을 반환
+    return ((toWeekStart.toEpochDays() - fromWeekStart.toEpochDays()) / 7).toInt()
 }
 
 fun getWeekDates(date: LocalDate, startFromMonday: Boolean = false): List<CalendarDate> {
